Modulize logging functions to loggingUser.js
