+++ js/init.js
@@ -1,3 +1,5 @@
+import { getUserData, isLogged } from './utils/loggingUser.js';
+
 const navLinks = document.querySelectorAll('.nav-item');
 
 let showSpinner = function () {
@@ -20,50 +22,6 @@ document.addEventListener('DOMContentLoaded', () => {
   createUserNav();
 });
 
-function isLogged() {
-  const userData = JSON.parse(localStorage.getItem('userData'));
-  return userData !== null && userData !== undefined;
-}
-
-// Logging user
-function login(userData) {
-  if (isLogged()) {
-    // console.log('El usuario ya esta logeado');
-    return false;
-  }
-
-  if (userData === null) {
-    // console.log('user Data es nulo');
-    return false;
-  }
-
-  if (!('email' in userData) || !('password' in userData)) {
-    // console.log(
-    // 'No existe el valor de email y contraseña en los datos del usuario'
-    // );
-    return false;
-  }
-
-  localStorage.setItem('userData', JSON.stringify(userData));
-  return true;
-}
-
-function logout() {
-  if (!isLogged()) {
-    return false;
-  }
-
-  localStorage.removeItem('userData');
-  return true;
-}
-
-function getUserData() {
-  if (!isLogged()) {
-    return null;
-  }
-  return JSON.parse(localStorage.getItem('userData'));
-}
-
 function getCurrentPage() {
   const pathname = window.location.pathname;
   const split = pathname.split('/');
@@ -76,8 +34,10 @@ function checkPage(target) {
 
 // Creates the UI User Nav
 const createUserNav = () => {
-  const { email } = getUserData();
-  navLinks[
-    navLinks.length - 1
-  ].innerHTML = `<span class="nav-link" id="loggeado">${email}</span>`;
+  if (getUserData() != null) {
+    const { email } = getUserData();
+    navLinks[
+      navLinks.length - 1
+    ].innerHTML = `<span class="nav-link" id="loggeado">${email}</span>`;
+  }
 };
